+++ packages/node/src/configure/NodeConfig.ts
@@ -1,0 +1,30 @@
+// Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
+// SPDX-License-Identifier: GPL-3.0
+
+import { IConfig, NodeConfig } from '@subql/node-core';
+
+export interface ISubstrateConfig extends IConfig {
+  skipTransactions: boolean;
+}
+
+export class SubstrateNodeConfig extends NodeConfig<ISubstrateConfig> {
+  /**
+   * This is a wrapper around the core NodeConfig to get additional properties that are provided through args or node runner options
+   * NOTE: This isn't injected anywhere so you need to wrap the injected node config
+   *
+   * @example
+   * constructor(
+   *   nodeConfig: NodeConfig,
+   * ) {
+   *   this.nodeConfig = new SubstrateNodeConfig(nodeConfig);
+   * }
+   * */
+  constructor(config: NodeConfig) {
+    // Rebuild with internal config
+    super((config as any)._config, (config as any)._isTest);
+  }
+
+  get skipTransactions(): boolean {
+    return !!this._config.skipTransactions;
+  }
+}