+++ packages/types-core/src/project/versioned/v1_0_0/types.ts
@@ -108,10 +108,11 @@ export interface NodeOptions {
   unfinalizedBlocks?: boolean;
 
   /**
-   * Indicates whether block skipping is enabled for the indexer (optional).
+   * Indicates whether transaction skipping is enabled for the indexer. (optional).
+   * If this is enabled and the project only contains event handlers then it wont fetch transactions reducing memory footprint and RPC requests. This can lead to faster indexing.
    * @type {boolean}
    */
-  skipBlock?: boolean;
+  skipTransactions?: boolean;
 }
 
 /**